Added rir utils, refactored reflections thin out
